Refactor: move topicDetailView to pages and rename to TopicView
